modules/cap-ci-cd-utils/requirements.txt
GitPython
Repo
google-cloud-storage==1.28.0
jsonpath-ng
requests